requirements.txt
streamlit==1.22.0
Pillow>=10.0.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==0.8.11
gTTS==2.3.1
bcrypt==4.0.1
itsdangerous==2.1.2
stripe==5.4.0
SQLAlchemy==2.0.15
pytest==7.3.1
pydantic==1.10.8
python-multipart==0.0.6
email-validator==2.0.0
python-dotenv==1.0.0
jinja2==3.1.2
weasyprint==59.0
